docs: Make line-limited literalincludes less fragile
Particularly, harden them against code motion.

Use :end-before: instead of line numbers to stop the full program
listings before the test routines.  To this end, add a sentinel
comment line to the python programs where the literalinclude should
end.

Use the :pyobject: line-selection helper in the one case it applies --
the pure python quant().  It can't be used for the Chapel version
because it won't include the leading @Chapel().

I didn't use :begin-after: and :end-before: sentinels for the
@Chapel() quant either because they'd be included in the full-program
listing.